memory_agent/requirements.txt
langchain
langchain-openai
langchain-google-genai
qdrant-client
fastapi
uvicorn
python-dotenv
streamlit
pydantic
langchain-huggingface